change type to double

+++ Distances.cpp
@@ -5,26 +5,26 @@
 
 using namespace std;
 // constructor
-Distances ::Distances(vector<long double> &v1, const vector<long double> &v2)
+Distances ::Distances(vector<double> &v1, const vector<double> &v2)
 {
     this->v1 = v1;
     this->v2 = v2;
 }
 // calculate euclidian distance
-long double Distances::euclidianDistance()
+double Distances::euclidianDistance()
 {
     return minkowskiDistanceGeneric(2);
 }
 // calculate manhatten distance
-long double Distances::manhattenDistance()
+double Distances::manhattenDistance()
 {
     return minkowskiDistanceGeneric(1);
 }
 
 //calculate chebyshev distance
-long double Distances::chebyshevDistance()
+double Distances::chebyshevDistance()
 {
-    long double distance = 0;
+    double distance = 0;
     for (int i = 0; i < v1.size(); ++i)
     {
         if (abs(v1[i] - v2[i]) > distance)
@@ -33,23 +33,23 @@ long double Distances::chebyshevDistance()
     return distance;
 }
 // calculate canberra distance
-long double Distances::canberraDistance()
+double Distances::canberraDistance()
 {
-    long double distance = 0;
+    double distance = 0;
     for (int i = 0; i < v1.size(); ++i)
     {
-        distance += (long double)abs(v1[i] - v2[i]) / (long double)(abs(v1[i]) + abs(v2[i]));
+        distance += (double)abs(v1[i] - v2[i]) / (double)(abs(v1[i]) + abs(v2[i]));
     }
     return distance;
 }
 // calculate minkowski distance
-long double Distances::minkowskiDistance()
+double Distances::minkowskiDistance()
 {
     return minkowskiDistanceGeneric(P);
 }
-long double Distances::minkowskiDistanceGeneric(int exp)
+double Distances::minkowskiDistanceGeneric(int exp)
 {
-    long double distance = 0;
+    double distance = 0;
     for (int i = 0; i < v1.size(); ++i)
     {
         distance += pow(abs(v1[i] - v2[i]), exp);
